requirements.txt
-i https://pypi.org/simple
alembic==1.12.0 ; python_version >= '3.7'
bcrypt==4.0.1 ; python_version >= '3.6'
blinker==1.6.3 ; python_version >= '3.7'
click==8.1.7 ; python_version >= '3.7'
flask==3.0.0
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-migrate==4.0.5
flask-sqlalchemy==3.1.1
greenlet==3.0.0 ; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==21.2.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
mako==1.2.4 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
packaging==23.2 ; python_version >= '3.7'
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlalchemy==2.0.22
sqlalchemy-serializer==1.4.1
typing-extensions==4.8.0 ; python_version >= '3.8'
werkzeug==3.0.0 ; python_version >= '3.8'
zipp==3.17.0 ; python_version >= '3.8'
